requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
codegen==1.0
cryptography==2.2.2
diff-match-patch==20121119
Django==2.0.4
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.6
et-xmlfile==1.0.1
Fabric3==1.14.post1
future==0.16.0
httplib2==0.9.2
idna==2.6
jdcal==1.4
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.2
paramiko==2.4.1
patterns==0.3
Pillow==5.1.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
pytz==2018.4
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
